Changing real(8) to real(real64) for better portability

+++ Pragma_Examples/OpenMP/Fortran/3_vecadd/1_vecadd_usm/vecadd.f90
@@ -3,16 +3,17 @@
 program main
 
     use omp_lib 
+    use, intrinsic :: iso_fortran_env, only: real64
     implicit none
     !$omp requires unified_shared_memory
     ! Size of vectors
     integer :: n = 10000000
  
     ! Input vectors and Output vector
-    real(8),dimension(:),allocatable :: a, b, c
+    real(real64),dimension(:),allocatable :: a, b, c
     integer :: i
-    real(8) :: sum
-    real(8) :: startt, endt
+    real(real64) :: sum
+    real(real64) :: startt, endt
 
     ! Allocate memory for each vector
     allocate(a(n), b(n), c(n))
